+++ src/atoms/ButtonQR/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-import ButtonQR from './ButtonQR';
-
-export default ButtonQR;