+++ src/components/Invoice/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import Invoice from './Invoice';
+
+export default Invoice;