+++ src/components/index.js
@@ -1,5 +1,4 @@
 // @flow
 
-export { default as BadgerButton } from './BadgerButton';
-export { default as BadgerBadge } from './BadgerBadge';
+export { default as Invoice } from './Invoice';
 export { default as PriceDisplay } from './PriceDisplay';